解决dlmalloc分配失败问题

+++ test/memory/buddy.cpp
@@ -25,7 +25,7 @@ int main(int argc, char *argv[]) {
             }
         }
         if (cmd == 0) {
-            if (v.empty() || size >= v.size())
+            if (v.empty() || size >= (int)v.size())
                 continue;
             vector<void*>::iterator it = v.begin();
             void *p = *(it + size);
